% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{bedasbigmatrix}
\alias{bedasbigmatrix}
\title{Get Bed Data As A BigMatrix}
\usage{
bedasbigmatrix(
  bed,
  filename = NULL,
  meanimpute = FALSE,
  normalise = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{bed}{The "rbed" object to read}

\item{filename}{(default NULL, meaning use in-memory representation) The location of a file to use to store the matrix}

\item{meanimpute}{(default FALSE) Whether to replace missing values with their mean (if normalise=TRUE only)}

\item{normalise}{(default TRUE) Whether to normalise data by SNP frequency, or return counts.}

\item{verbose}{(default TRUE) whether to show progress for individual access}
}
\value{
A bigmatrix of
}
\description{
Gets a bed file as a BigMatrix object from package "big.memory".
}
\seealso{
\code{\link{bedasbigmatrix}}
}
