% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{bedashap}
\alias{bedashap}
\title{Convert Rbed object to binary haplotype format}
\usage{
bedashap(dat, ploidy = 2)
}
\arguments{
\item{dat}{an rbed object as output by \code{\link{readbed}}}

\item{ploidy}{the ploidy of the data. The only plausible value for plink format is 2.}
}
\value{
a "hapmatrix" object, which is a list containing:
\itemize{
\item data: The SNP matrix in haplotype format with pN rows and L columns (p=ploidy)
\item snpinfo: as provided from the rbed object
\item indinfo: as provided from the rbed object
\item hapinfo: an N by p matrix giving individual identifiers for the haplotypes
}
}
\description{
Convert an rbed object from plink's native file format into a haplotype-based binary matrix. This involves pseudo-phasing, where alleles are assigned at random to either of the two haplotypes.
}
