% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{bedasmatrix}
\alias{bedasmatrix}
\title{Get Bed Data As A Matrix}
\usage{
bedasmatrix(bed, meanimpute = FALSE, normalise = TRUE, verbose = TRUE)
}
\arguments{
\item{bed}{The "rbed" object to read}

\item{meanimpute}{(default FALSE) Whether to replace missing values with their mean (if normalise=TRUE only)}

\item{normalise}{(default TRUE) Whether to normalise data by SNP frequency, or return counts.}

\item{verbose}{(default TRUE) whether to show progress for individual access}
}
\value{
A matrix of data with the additional attribute "idx" referring to the indexes of the returned individuals.
}
\description{
Gets a bed file as a regular R matrix.
}
\seealso{
\code{\link{bedasbigmatrix}}
}
