% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{get_data}
\alias{get_data}
\title{Get normalised data from an rbed object in an efficient format}
\usage{
get_data(bed, i, meanimpute = FALSE, normalise = TRUE, verbose = TRUE)
}
\arguments{
\item{bed}{The "rbed" object to read}

\item{i}{The index of data to read. 4 individuals are stored in each binary object. so i refers to the index of this set, i.e. i=2 gives individuals 5:8. You can give a single value or a vector. It returns fewer entries if you are at the end of the data.}

\item{meanimpute}{default FALSE. If normalise=TRUE, whether to replace missing values with their mean.}

\item{normalise}{default TRUE. Whether to return the raw data or normalise by frequency.}

\item{verbose}{(default TREE) whether to show progress for individual access}
}
\value{
A matrix of data with the additional attribute "idx" referring to the indexes of the returned individuals.
}
\description{
This function is for efficient bulk processing. It accesses data in the format in which it is stored, rather than by individual IDs.
}
