% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbed.R
\name{getfreqs}
\alias{getfreqs}
\title{Compute the allele frequencies for bed file}
\usage{
getfreqs(bed, verbose = TRUE)
}
\arguments{
\item{bed}{An "rbed" object as returned from \code{\link{readbed}} or \code{\link{mergeref}}}

\item{verbose}{(default TRUE) whether to report on activities}
}
\value{
The bed object provided with the $freq
}
\description{
Adds the SNP frequency calculations to an "rbed" object.
}
