% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbed.R
\name{getind}
\alias{getind}
\title{Extract a individual from a binary format bed object}
\usage{
getind(bed, x)
}
\arguments{
\item{bed}{An "rbed" object as returned from \code{\link{readbed}} or \code{\link{mergeref}}}

\item{x}{The index of the individual to access}
}
\value{
a vector of length (bed$no.snp or length(bed$datkeep)) containing the SNP values for each individual
}
\description{
The binary bed format stores 4 individuals in a byte. This returns the SNP details for these one individual only. It is more efficient to get the chunks with \code{\link{getinds}}.
}
\seealso{
\code{\link{getinds}}, \code{\link{getsnp}}
}
