% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbed.R
\name{getsnp}
\alias{getsnp}
\title{Extract a SNP from a binary format bed object}
\usage{
getsnp(bed, x)
}
\arguments{
\item{bed}{An "rbed" object as returned from \code{\link{readbed}} or \code{\link{mergeref}}}

\item{x}{The index of the SNP to extract}
}
\value{
a vector of length bed$no.ind containing the SNP values for each individual
}
\description{
Convert the raw BED data into an integer string, apply any allele flipping inferred from the reference
}
\seealso{
\code{\link{getind}}, \code{\link{getinds}}
}
