% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbed.R
\name{mergeref}
\alias{mergeref}
\title{Merge your dataset into a reference using the efficient binary "rbed" format}
\usage{
mergeref(
  dat,
  ref = readreference(pcapred.ref::ukb_pcs_18()),
  mergeon = "ID",
  verbose = TRUE
)
}
\arguments{
\item{dat}{An object of class "rawdata" as returned by \code{\link{readinput}}.}

\item{ref}{(default: the provided \code{\link[pcapred.ref]{ukb_pcs_18}}) The reference dataset of class "referencedata" with which you want to merge your data. This is as returned by \code{link{readreference}}.}

\item{mergeon}{(default ID) how to merge the SNPs. "ID" is always available but assumes you have rsids for the data. Alternative is "CHRPOS" which uses the extended freq information we provide in the reference dataset.}

\item{verbose}{whether to say which algorithm is being used}
}
\value{
An object of class "mergeddata" which is a list containing:
\itemize{
\item snpstats (as in ref, see \code{link{readreference}})
\item snpweights (as in ref, see \code{link{readreference}})
\item pcvals (as in ref, see \code{link{readreference}})
\item mode  (as in ref, see \code{link{readreference}})
\item transpose (as in ref, see \code{link{readreference}})
\item minfreq (as applied, taken from parameters or ref)
\item refkeep A vector of indices for the retained SNPs in the reference
\item datkeep A vector of indices for the retained SNPs in the target dataset
\item flip A boolean vector of length = length(datkeep) for whether the alleles of that SNP need flipping
}
}
\description{
Harmonise your dataset with a reference, performing sanity checks and standardization of REF/ALT alleles.
}
