% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingfunctions.R
\name{normalise}
\alias{normalise}
\title{Normalise SNPs according to a reference frequency under a binomial model}
\usage{
normalise(rawdatmat, f, ploidy = 2, minfreq = NULL, meanimpute = FALSE)
}
\arguments{
\item{rawdatmat}{A matrix of SNP counts of the alternative allele; of dimension N by L}

\item{f}{A snp frequency vector of length L}

\item{ploidy}{ploidy of the data (2=diploid) ; maximum that rawdatmat can take}

\item{minfreq}{a thresholding for very raw SNPs in the reference, to prevent them from being over-weighted}

\item{meanimpute}{(default FALSE) whether to mean impute missing data}
}
\value{
A matrix of size N by L which is rawdatmat zero meaned and with theoretical s.d. 1
}
\description{
Normalise a SNP count matrix by an external reference frequency distribution by mean and standard deviation predicted under a binomial model.
}
