% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingfunctions.R
\name{pred_flash_missing}
\alias{pred_flash_missing}
\title{Create predictions for eigenvalues from flashpca-scaled data}
\usage{
pred_flash_missing(keepdat, ndat, res, L)
}
\arguments{
\item{keepdat}{a matrix of presence indicators}

\item{ndat}{normalised data matrix}

\item{res}{snp weighting matrix}

\item{L}{The number of non-missing SNPs}
}
\value{
A matrix of size N by K
}
\description{
Make a genome-wide prediction based on plink scalings
}
