% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingfunctions.R
\name{pred_missing}
\alias{pred_missing}
\title{Create predictions for eigenvalues when there is missing data}
\usage{
pred_missing(ndat, res, mode)
}
\arguments{
\item{ndat}{normalised data matrix}

\item{res}{snp weighting matrix}

\item{mode}{either "flashpca" or "plink"}
}
\value{
A matrix of size N by K
}
\description{
Make a genome-wide prediction based on incomplete data
}
