% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingfunctions.R
\name{pred_simple}
\alias{pred_simple}
\title{Create predictions for eigenvalues when there is no missing data}
\usage{
pred_simple(ndat, res, L, mode)
}
\arguments{
\item{ndat}{normalised data matrix}

\item{res}{snp weighting matrix}

\item{L}{the number of SNPs}

\item{mode}{either "flashpca" or "plink"}
}
\value{
A matrix of size N by K
}
\description{
Make a genome-wide prediction based on complete data, which makes the computation faster
}
