% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbed.R
\name{predictpcs}
\alias{predictpcs}
\title{Create predictions for eigenvalues from binary stored data}
\usage{
predictpcs(dat, minfreq = NULL, verbose = TRUE)
}
\arguments{
\item{dat}{An object of class "mergedrbed" as returned by \code{\link{mergeref}}}

\item{minfreq}{The minimum reference frequency to allow, to prevent rare SNPs having a disproportionate effect}

\item{verbose}{whether to say which algorithm is being used}
}
\value{
A matrix of size N by K of predicted Eigenvectors
}
\description{
Make a genome-wide prediction based on incomplete data
}
