% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{readinput}
\alias{readinput}
\title{Read the target dataset}
\usage{
readinput(inputfile, raw = FALSE, verbose = TRUE)
}
\arguments{
\item{inputfile}{file root to read}

\item{raw}{if TRUE, treat the input as a text file as output by plink --convert 12. Otherwise (default) read the .bim/.bed/.fam}

\item{verbose}{whether to report file reading progress}
}
\value{
See readbed
}
\description{
Read a bim/bed
}
