% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{readpcs}
\alias{readpcs}
\title{Read the target dataset}
\usage{
readpcs(pcvalsfile, ...)
}
\arguments{
\item{pcvalsfile}{file to read in}

\item{...}{extra parameters to read.table}
}
\value{
A matrix of the PCs
}
\description{
Read a bim/bed
}
