% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{readreference}
\alias{readreference}
\title{Read the Reference Data for a PC projection}
\usage{
readreference(
  fileroot,
  mode = "flashpca",
  transpose = FALSE,
  checkonly = FALSE,
  minfreq = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{fileroot}{fileset to read in}

\item{mode}{Data mode; one of flashpca (default) or plink}

\item{transpose}{Whether SNP weights are transposed (default FALSE)}

\item{checkonly}{default FALSE. If TRUE, no file reading is done but the files are checked for existing}

\item{minfreq}{(default 0.001) minimum allele frequency allowed in the normalisation phase, to prevent ultra-rare SNPs from being upweighted too much}

\item{verbose}{default TRUE. Whether to output progress.}
}
\value{
An object of class "referencedata" which is a list containing:
\itemize{
\item snpweights (if !checkonly) The matrix of snp weights, with SNPs as rows and PCs as columns
\item snpstats (if !checkonly) The matrix of snp statistics, with SNPs as rows
\item pcvals (if !checkonly) A vector of PC Eigenvalues (plink) or SVD Singular vectors (flashpca) as appropriate.
\item mode As provided in the call.
\item transpose As provided in the call
\item minfreq As provided in the call
\item pcvalsfile The file intended to be used for pcvals
\item snpweightsfile The file intended to be used for snpweights
\item snpstatsfile The file intended to be used for snpstats
}
}
\description{
Read in SNP weightings, eigenvalues and SNP info.


SNP info must be in <fileroot>.afreq.gz
Weights are read according to mode and transpose flag.

Plink weights are in <fileroot>.eigenvec.var.gz/<fileroot>.eigenvec.tvar.gz
flashpca weights are in <fileroot>.tload.gz/<fileroot>.tload.gz

Plink (Eigen) values are in <fileroot>.eigenval
flashpca (Singular) values are in <fileroot>.val
}
\examples{
\dontrun{
tfile=gsub(".load.gz","",system.file("extdata","ukb_pcs_18.load.gz",package="pcapred"))
ref=readreference(tfile)
}
}
