% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{readsnpstats}
\alias{readsnpstats}
\title{Read the snp stats}
\usage{
readsnpstats(afreqfile, ...)
}
\arguments{
\item{afreqfile}{file to read in}

\item{...}{extra parameters to read.table}
}
\value{
A matrix of the SNP stats
}
\description{
Read a afreq file from plink
}
