% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{readsnpweights}
\alias{readsnpweights}
\title{Read Snp Weights/Loadings from a file}
\usage{
readsnpweights(
  snpweightsfile,
  mode = "flashpca",
  transpose = FALSE,
  colnames = NULL,
  verbose = TRUE
)
}
\arguments{
\item{snpweightsfile}{The file to read}

\item{mode}{either flashpca or plink; default "flashpca"}

\item{transpose}{default FALSE, meaning treat columns as PCs and rows as SNPs. if TRUE, assume the laoding file is transposed from this.}

\item{colnames}{(default NULL, meaning do no naming) names for the SNPs' comes from}

\item{verbose}{(default TRuE) whether to show progress for file reading}
}
\value{
a matrix of SNP weights with K rows and L columns
}
\description{
Read SNP weights from a variety of formats (plink/flashpca, either regular or transposed)
}
