% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingfunctions.R
\name{transform_flashref}
\alias{transform_flashref}
\title{Create direct SNP weights from a flashpca-like (scaled) reference}
\usage{
transform_flashref(snpweights, ...)
}
\arguments{
\item{snpweights}{A weighting matrix of dimension L by K}

\item{...}{Ignored parameters (for using the same call structure as other transforms)}
}
\value{
a canonical scaled version of the snpweights matrix
}
\description{
PCA can be represented in an a number of different ways, provided that the structure
$$Y = U D V^T$$
flashpca uses this canonical form but the weights are out by a factor \eqn{\sqrt{2}} due to diploid genomes.
}
\seealso{
transform.plinkref
}
