% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingfunctions.R
\name{transform_pcs}
\alias{transform_pcs}
\title{Create direct SNP weights from a reference}
\usage{
transform_pcs(snpweights, pcvals, mode)
}
\arguments{
\item{snpweights}{A weighting matrix of dimension L by K}

\item{pcvals}{Eigenvalues from the original PCA}

\item{mode}{"plink" pr "flashpca"}
}
\value{
a canonical scaled version of the snpweights matrix
}
\description{
PCA can be represented in an a number of different ways, provided that the structure
$$Y = U D V^T$$
flashpca uses this canonical form but the weights are out by a factor \eqn{\sqrt{2}} due to diploid genomes. Plink-like loadings instead scale U by the square root of the eigenvalues.
}
\seealso{
transform.plinkref
}
