% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingfunctions.R
\name{transform_plinkref}
\alias{transform_plinkref}
\title{Create direct SNP weights from a plink-like (scaled) reference}
\usage{
transform_plinkref(snpweights, lambda)
}
\arguments{
\item{snpweights}{A weighting matrix of dimension L by K.}

\item{lambda}{A n eigenvalue vector of length K}
}
\value{
a canonical scaled version of the snpweights matrix
}
\description{
PCA can be represented in an a number of different ways, provided that the structure
$$Y = U D V^T$$
holds, where $D$ is a diagonal matrix of the Eigenvalues. In particular, plink and FastPCA use \eqn{U' = U \sqrt{D}} using $Y=U'{V'}^T$.

This function converts $V'$ into the canonical form, which also corrects for the factor \eqn{\sqrt{2}} from diploid genomes.
}
