% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{writepred}
\alias{writepred}
\title{Write Predicted PCs In Standard Format}
\usage{
writepred(outfile, dat, pred)
}
\arguments{
\item{outfile}{file to write to}

\item{dat}{object of class "mergeddata" as returned by \code{\link{mergeref}}, or a dataframe with "FamID" amd "IndID" columns}

\item{pred}{The predicted PCs}
}
\value{
NULL, invisibly
}
\description{
Output the predicted PCs into a file, in the plink covariate format.
}
